backend/health/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
colorama==0.4.6
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysqlclient==2.2.7
openai==1.63.2
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
